Switch from _makefile_refs to _io_refs
The socket class needs to track references to the file created with
`makefile`, in order to know if the underlying can be closed safely.
PyPy 2 uses `_makefile_refs` for this, while PyPy 3 and CPython share
the same code and use `_io_refs` instead. Since we no longer support
Python 2, we can drop the code specific to PyPy 2 and switch to
`_io_refs`. This will allow to remove our makefile backport that used
`_makefile_refs` for PyPy 2 compatibility until now.