Import find_unused_port from standard library
The idea is to use a newer version of `find_unused_port` that hopefully
is faster on Windows.

This means I had to rename the tests directory, as CPython also uses
"test". If this experiment is successful, I will update the copy/paste
as the test module is not designed for external consumption anyway.